+++ server/private/routers/alertRule/listAlertRules.ts
@@ -84,28 +84,6 @@ const HEALTH_CHECK_ALERT_EVENT_TYPES = [
     "health_check_toggle"
 ] as const;
 
-export type ListAlertRulesResponse = {
-    alertRules: {
-        alertRuleId: number;
-        orgId: string;
-        name: string;
-        eventType: string;
-        enabled: boolean;
-        cooldownSeconds: number;
-        lastTriggeredAt: number | null;
-        createdAt: number;
-        updatedAt: number;
-        siteIds: number[];
-        healthCheckIds: number[];
-        resourceIds: number[];
-    }[];
-    pagination: {
-        total: number;
-        limit: number;
-        offset: number;
-    };
-};
-
 registry.registerPath({
     method: "get",
     path: "/org/{orgId}/alert-rules",
